fix(setup bug fix)
