--- training/asr_trainings/asr_model_taining_2.py
@@ -1,156 +1,0 @@
-import os
-from datasets import load_dataset, DatasetDict, Audio
-from transformers import WhisperFeatureExtractor, WhisperProcessor, WhisperForConditionalGeneration, Seq2SeqTrainingArguments, Seq2SeqTrainer
-from huggingface_hub import login
-from transformers import WhisperTokenizer
-
-import torch
-
-from dataclasses import dataclass
-from typing import Any, Dict, List, Union
-import evaluate
-
-# Define the model name at the top-level
-MODEL_NAME = "./whisper_finetuned_marathi_paani"
-
-# Hugging Face login using your token (set this in the environment variable 'HUGGINGFACE_TOKEN')
-huggingface_token = os.getenv('HUGGINGFACE_TOKEN')
-login(token=huggingface_token)
-
-os.environ["PYTORCH_MPS_HIGH_WATERMARK_RATIO"] = "0.0"
-
-common_voice = DatasetDict()
-common_voice["train"] = load_dataset("mozilla-foundation/common_voice_11_0", "mr", split="train+validation")
-common_voice["test"] = load_dataset("mozilla-foundation/common_voice_11_0", "mr", split="test")
-common_voice = common_voice.remove_columns(["accent", "age", "client_id", "down_votes", "gender", "locale", "path", "segment", "up_votes"])
-
-print(common_voice)
-
-feature_extractor = WhisperFeatureExtractor.from_pretrained("openai/whisper-small")
-
-tokenizer = WhisperTokenizer.from_pretrained("openai/whisper-small", language="Marathi", task="transcribe")
-
-input_str = common_voice["train"][0]["sentence"]
-labels = tokenizer(input_str).input_ids
-decoded_with_special = tokenizer.decode(labels, skip_special_tokens=False)
-decoded_str = tokenizer.decode(labels, skip_special_tokens=True)
-
-print(f"Input:                 {input_str}")
-print(f"Decoded w/ special:    {decoded_with_special}")
-print(f"Decoded w/out special: {decoded_str}")
-print(f"Are equal:             {input_str == decoded_str}")
-
-
-processor = WhisperProcessor.from_pretrained("openai/whisper-small", language="Marathi", task="transcribe")
-
-print(common_voice["train"][0])
-
-common_voice = common_voice.cast_column("audio", Audio(sampling_rate=16000))
-
-print(common_voice["train"][0])
-
-def prepare_dataset(batch):
-    # load and resample audio data from 48 to 16kHz
-    audio = batch["audio"]
-
-    # compute log-Mel input features from input audio array 
-    batch["input_features"] = feature_extractor(audio["array"], sampling_rate=audio["sampling_rate"]).input_features[0]
-
-    # encode target text to label ids 
-    batch["labels"] = tokenizer(batch["sentence"]).input_ids
-    return batch
-
-common_voice = common_voice.map(prepare_dataset, remove_columns=common_voice.column_names["train"], num_proc=4)
-
-model = WhisperForConditionalGeneration.from_pretrained("openai/whisper-small")
-
-model.generation_config.language = "marathi"
-model.generation_config.task = "transcribe"
-model.config.gradient_checkpointing = True
-
-model.generation_config.forced_decoder_ids = None
-
-@dataclass
-class DataCollatorSpeechSeq2SeqWithPadding:
-    processor: Any
-    decoder_start_token_id: int
-
-    def __call__(self, features: List[Dict[str, Union[List[int], torch.Tensor]]]) -> Dict[str, torch.Tensor]:
-        # split inputs and labels since they have to be of different lengths and need different padding methods
-        # first treat the audio inputs by simply returning torch tensors
-        input_features = [{"input_features": feature["input_features"]} for feature in features]
-        batch = self.processor.feature_extractor.pad(input_features, return_tensors="pt")
-
-        # get the tokenized label sequences
-        label_features = [{"input_ids": feature["labels"]} for feature in features]
-        # pad the labels to max length
-        labels_batch = self.processor.tokenizer.pad(label_features, return_tensors="pt")
-
-        # replace padding with -100 to ignore loss correctly
-        labels = labels_batch["input_ids"].masked_fill(labels_batch.attention_mask.ne(1), -100)
-
-        # if bos token is appended in previous tokenization step,
-        # cut bos token here as it's append later anyways
-        if (labels[:, 0] == self.decoder_start_token_id).all().cpu().item():
-            labels = labels[:, 1:]
-
-        batch["labels"] = labels
-
-        return batch
-    
-data_collator = DataCollatorSpeechSeq2SeqWithPadding(
-    processor=processor,
-    decoder_start_token_id=model.config.decoder_start_token_id,
-)
-
-metric = evaluate.load("wer")
-
-def compute_metrics(pred):
-    pred_ids = pred.predictions
-    label_ids = pred.label_ids
-
-    # replace -100 with the pad_token_id
-    label_ids[label_ids == -100] = tokenizer.pad_token_id
-
-    # we do not want to group tokens when computing the metrics
-    pred_str = tokenizer.batch_decode(pred_ids, skip_special_tokens=True)
-    label_str = tokenizer.batch_decode(label_ids, skip_special_tokens=True)
-
-    wer = 100 * metric.compute(predictions=pred_str, references=label_str)
-
-    return {"wer": wer}
-
-training_args = Seq2SeqTrainingArguments(
-    output_dir="./whisper-small-mr",  # change to a repo name of your choice
-    per_device_train_batch_size=4,
-    gradient_accumulation_steps=1,  # increase by 2x for every 2x decrease in batch size
-    learning_rate=1e-5,
-    warmup_steps=100,
-    max_steps=100,
-    gradient_checkpointing=True,
-    fp16=False,
-    eval_strategy="steps",
-    per_device_eval_batch_size=8,
-    predict_with_generate=True,
-    generation_max_length=225,
-    save_steps=100,
-    eval_steps=100,
-    logging_steps=25,
-    report_to=["tensorboard"],
-    load_best_model_at_end=True,
-    metric_for_best_model="wer",
-    greater_is_better=False,
-    push_to_hub=False,
-)
-
-trainer = Seq2SeqTrainer(
-    args=training_args,
-    model=model,
-    train_dataset=common_voice["train"],
-    eval_dataset=common_voice["test"],
-    data_collator=data_collator,
-    compute_metrics=compute_metrics,
-    tokenizer=processor.feature_extractor,
-)
-
-trainer.train()
